fix/perf(forge): improve glob path matching
