+++ cli/src/cmd/forge/test/mod.rs
@@ -7,10 +7,9 @@ use crate::{
     compile,
     compile::ProjectCompiler,
     suggestions, utils,
-    utils::FoundryPathExt,
 };
 use clap::{AppSettings, Parser};
-use ethers::solc::{utils::RuntimeOrHandle, FileFilter};
+use ethers::solc::utils::RuntimeOrHandle;
 use forge::{
     decode::decode_console_logs,
     executor::{inspector::CheatsConfig, opts::EvmOpts},
@@ -20,200 +19,17 @@ use forge::{
         identifier::{EtherscanIdentifier, LocalTraceIdentifier},
         CallTraceDecoderBuilder, TraceKind,
     },
-    MultiContractRunner, MultiContractRunnerBuilder, TestFilter,
+    MultiContractRunner, MultiContractRunnerBuilder,
 };
 use foundry_common::evm::EvmArgs;
 use foundry_config::{figment::Figment, Config};
 use regex::Regex;
-use std::{
-    collections::BTreeMap,
-    fmt,
-    path::{Path, PathBuf},
-    sync::mpsc::channel,
-    thread,
-    time::Duration,
-};
+use std::{collections::BTreeMap, path::PathBuf, sync::mpsc::channel, thread, time::Duration};
 use tracing::trace;
 use watchexec::config::{InitConfig, RuntimeConfig};
 use yansi::Paint;
-
-#[derive(Debug, Clone, Parser)]
-pub struct Filter {
-    /// Only run test functions matching the specified regex pattern.
-    ///
-    /// Deprecated: See --match-test
-    #[clap(long = "match", short = 'm')]
-    pub pattern: Option<regex::Regex>,
-
-    /// Only run test functions matching the specified regex pattern.
-    #[clap(
-        long = "match-test",
-        visible_alias = "mt",
-        conflicts_with = "pattern",
-        value_name = "REGEX"
-    )]
-    pub test_pattern: Option<regex::Regex>,
-
-    /// Only run test functions that do not match the specified regex pattern.
-    #[clap(
-        long = "no-match-test",
-        visible_alias = "nmt",
-        conflicts_with = "pattern",
-        value_name = "REGEX"
-    )]
-    pub test_pattern_inverse: Option<regex::Regex>,
-
-    /// Only run tests in contracts matching the specified regex pattern.
-    #[clap(
-        long = "match-contract",
-        visible_alias = "mc",
-        conflicts_with = "pattern",
-        value_name = "REGEX"
-    )]
-    pub contract_pattern: Option<regex::Regex>,
-
-    /// Only run tests in contracts that do not match the specified regex pattern.
-    #[clap(
-        long = "no-match-contract",
-        visible_alias = "nmc",
-        conflicts_with = "pattern",
-        value_name = "REGEX"
-    )]
-    pub contract_pattern_inverse: Option<regex::Regex>,
-
-    /// Only run tests in source files matching the specified glob pattern.
-    #[clap(
-        long = "match-path",
-        visible_alias = "mp",
-        conflicts_with = "pattern",
-        value_name = "GLOB"
-    )]
-    pub path_pattern: Option<globset::Glob>,
-
-    /// Only run tests in source files that do not match the specified glob pattern.
-    #[clap(
-        name = "no-match-path",
-        long = "no-match-path",
-        visible_alias = "nmp",
-        conflicts_with = "pattern",
-        value_name = "GLOB"
-    )]
-    pub path_pattern_inverse: Option<globset::Glob>,
-}
-
-impl Filter {
-    pub fn with_merged_config(&self, config: &Config) -> Self {
-        let mut filter = self.clone();
-        if filter.test_pattern.is_none() {
-            filter.test_pattern = config.test_pattern.clone().map(|p| p.into());
-        }
-        if filter.test_pattern_inverse.is_none() {
-            filter.test_pattern_inverse = config.test_pattern_inverse.clone().map(|p| p.into());
-        }
-        if filter.contract_pattern.is_none() {
-            filter.contract_pattern = config.contract_pattern.clone().map(|p| p.into());
-        }
-        if filter.contract_pattern_inverse.is_none() {
-            filter.contract_pattern_inverse =
-                config.contract_pattern_inverse.clone().map(|p| p.into());
-        }
-        if filter.path_pattern.is_none() {
-            filter.path_pattern = config.path_pattern.clone();
-        }
-        if filter.path_pattern_inverse.is_none() {
-            filter.path_pattern_inverse = config.path_pattern_inverse.clone();
-        }
-        filter
-    }
-}
-
-impl FileFilter for Filter {
-    /// Returns true if the file regex pattern match the `file`
-    ///
-    /// If no file regex is set this returns true if the file ends with `.t.sol`, see
-    /// [FoundryPathExr::is_sol_test()]
-    fn is_match(&self, file: &Path) -> bool {
-        if let Some(file) = file.as_os_str().to_str() {
-            if let Some(ref glob) = self.path_pattern {
-                return glob.compile_matcher().is_match(file)
-            }
-            if let Some(ref glob) = self.path_pattern_inverse {
-                return !glob.compile_matcher().is_match(file)
-            }
-        }
-        file.is_sol_test()
-    }
-}
-
-impl TestFilter for Filter {
-    fn matches_test(&self, test_name: impl AsRef<str>) -> bool {
-        let mut ok = true;
-        let test_name = test_name.as_ref();
-        // Handle the deprecated option match
-        if let Some(re) = &self.pattern {
-            ok &= re.is_match(test_name);
-        }
-        if let Some(re) = &self.test_pattern {
-            ok &= re.is_match(test_name);
-        }
-        if let Some(re) = &self.test_pattern_inverse {
-            ok &= !re.is_match(test_name);
-        }
-        ok
-    }
-
-    fn matches_contract(&self, contract_name: impl AsRef<str>) -> bool {
-        let mut ok = true;
-        let contract_name = contract_name.as_ref();
-        if let Some(re) = &self.contract_pattern {
-            ok &= re.is_match(contract_name);
-        }
-        if let Some(re) = &self.contract_pattern_inverse {
-            ok &= !re.is_match(contract_name);
-        }
-        ok
-    }
-
-    fn matches_path(&self, path: impl AsRef<str>) -> bool {
-        let mut ok = true;
-        let path = path.as_ref();
-        if let Some(ref glob) = self.path_pattern {
-            ok &= glob.compile_matcher().is_match(path);
-        }
-        if let Some(ref glob) = self.path_pattern_inverse {
-            ok &= !glob.compile_matcher().is_match(path);
-        }
-        ok
-    }
-}
-
-impl fmt::Display for Filter {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let mut patterns = Vec::new();
-        if let Some(ref p) = self.pattern {
-            patterns.push(format!("\tmatch: `{}`", p.as_str()));
-        }
-        if let Some(ref p) = self.test_pattern {
-            patterns.push(format!("\tmatch-test: `{}`", p.as_str()));
-        }
-        if let Some(ref p) = self.test_pattern_inverse {
-            patterns.push(format!("\tno-match-test: `{}`", p.as_str()));
-        }
-        if let Some(ref p) = self.contract_pattern {
-            patterns.push(format!("\tmatch-contract: `{}`", p.as_str()));
-        }
-        if let Some(ref p) = self.contract_pattern_inverse {
-            patterns.push(format!("\tno-match-contract: `{}`", p.as_str()));
-        }
-        if let Some(ref p) = self.path_pattern {
-            patterns.push(format!("\tmatch-path: `{}`", p.glob()));
-        }
-        if let Some(ref p) = self.path_pattern_inverse {
-            patterns.push(format!("\tno-match-path: `{}`", p.glob()));
-        }
-        write!(f, "{}", patterns.join("\n"))
-    }
-}
+mod filter;
+pub use filter::Filter;
 
 // Loads project's figment and merges the build cli arguments into it
 foundry_config::impl_figment_convert!(TestArgs, opts, evm_opts);
@@ -472,6 +288,8 @@ pub fn custom_run(args: TestArgs, include_fuzz_tests: bool) -> eyre::Result<Test
     };
     let fuzzer = proptest::test_runner::TestRunner::new(cfg);
     let mut filter = args.filter(&config);
+
+    trace!(target: "forge::test", ?filter, "using filter");
 
     // Set up the project
     let project = config.project()?;
