+++ forge/src/multi_runner.rs
@@ -51,6 +51,7 @@ pub struct MultiContractRunner {
 }
 
 impl MultiContractRunner {
+    /// Returns the number of matching tests
     pub fn count_filtered_tests(&self, filter: &impl TestFilter) -> usize {
         self.contracts
             .iter()
